refactor: extract TinyOS browser app
